refactor: align public sdk import names
