+++ Cargo.toml
@@ -1,3 +1,8 @@
 [workspace]
 resolver = "2"
-members  = ["crates/mcp-compressor-core", "crates/mcp-compressor-python", "crates/mcp-compressor-node"]
+members  = [
+    "crates/mcp-compressor",
+    "crates/mcp-compressor-core",
+    "crates/mcp-compressor-python",
+    "crates/mcp-compressor-node",
+]
