+++ crates/mcp-compressor/src/lib.rs
@@ -1,0 +1,6 @@
+//! Public Rust SDK crate for mcp-compressor.
+//!
+//! This crate intentionally hides the internal `mcp-compressor-core` crate name
+//! from end-user Rust code while re-exporting the public SDK/runtime surface.
+
+pub use mcp_compressor_core::*;
