+++ python/mcp-compressor-rust/mcp_compressor/cli.py
