+++ python/mcp-compressor-rust/mcp_compressor/core.py
@@ -8,7 +8,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Any
 
-_mcp_compressor_core = importlib.import_module("mcp_compressor_rust._mcp_compressor_core")
+_native = importlib.import_module("mcp_compressor._native")
 
 
 @dataclass(frozen=True)
@@ -92,17 +92,17 @@ def _tool_payload(tools: list[ToolSpec]) -> str:
 
 def compress_tool_listing(level: str, tools: list[ToolSpec]) -> str:
     """Format a Rust-core compressed tool listing."""
-    return str(_mcp_compressor_core.compress_tool_listing_json(level, _tool_payload(tools)))
+    return str(_native.compress_tool_listing_json(level, _tool_payload(tools)))
 
 
 def format_tool_schema_response(tool: ToolSpec) -> str:
     """Format a Rust-core schema response for one tool."""
-    return str(_mcp_compressor_core.format_tool_schema_response_json(_json_dumps(tool.to_json_dict())))
+    return str(_native.format_tool_schema_response_json(_json_dumps(tool.to_json_dict())))
 
 
 def parse_tool_argv(tool: ToolSpec, argv: list[str]) -> dict[str, Any]:
     """Parse generated CLI argv for one tool through the Rust core parser."""
-    parsed = _mcp_compressor_core.parse_tool_argv_json(
+    parsed = _native.parse_tool_argv_json(
         _json_dumps(tool.to_json_dict()),
         _json_dumps(argv),
     )
@@ -118,7 +118,7 @@ def start_compressed_session(
     backends: list[BackendConfig],
 ) -> CompressedSession:
     """Start a Rust-backed compressed session and local proxy."""
-    native_session = _mcp_compressor_core.start_compressed_session_json(
+    native_session = _native.start_compressed_session_json(
         _json_dumps(config.to_json_dict()),
         _json_dumps([backend.to_json_dict() for backend in backends]),
     )
@@ -130,7 +130,7 @@ def start_compressed_session_from_mcp_config(
     mcp_config_json: str,
 ) -> CompressedSession:
     """Start a Rust-backed compressed session from MCP config JSON."""
-    native_session = _mcp_compressor_core.start_compressed_session_from_mcp_config_json(
+    native_session = _native.start_compressed_session_from_mcp_config_json(
         _json_dumps(config.to_json_dict()),
         mcp_config_json,
     )
@@ -155,7 +155,7 @@ def generate_client_artifacts(
         "output_dir": str(output_dir),
         "session_pid": session_pid,
     }
-    raw = json.loads(_mcp_compressor_core.generate_client_artifacts_json(kind, _json_dumps(config)))
+    raw = json.loads(_native.generate_client_artifacts_json(kind, _json_dumps(config)))
     if not isinstance(raw, list):
         msg = "Rust core generate_client_artifacts_json returned non-list JSON"
         raise TypeError(msg)
@@ -163,7 +163,7 @@ def generate_client_artifacts(
 
 
 def normalize_sdk_servers(servers: dict[str, Any]) -> list[BackendConfig]:
-    raw = json.loads(_mcp_compressor_core.normalize_servers_json(_json_dumps(servers)))
+    raw = json.loads(_native.normalize_servers_json(_json_dumps(servers)))
     if not isinstance(raw, list):
         msg = "Rust core normalize_servers_json returned non-list JSON"
         raise TypeError(msg)
@@ -177,7 +177,7 @@ def normalize_sdk_servers(servers: dict[str, Any]) -> list[BackendConfig]:
 
 def parse_mcp_config(config_json: str) -> list[dict[str, Any]]:
     """Parse MCP config JSON through the Rust core topology parser."""
-    value = json.loads(_mcp_compressor_core.parse_mcp_config_json(config_json))
+    value = json.loads(_native.parse_mcp_config_json(config_json))
     if not isinstance(value, list):
         msg = "Rust core parse_mcp_config_json returned non-list JSON"
         raise TypeError(msg)
@@ -186,7 +186,7 @@ def parse_mcp_config(config_json: str) -> list[dict[str, Any]]:
 
 def list_oauth_credentials() -> list[dict[str, Any]]:
     """List remembered Rust OAuth credential stores."""
-    value = json.loads(_mcp_compressor_core.list_oauth_credentials_json())
+    value = json.loads(_native.list_oauth_credentials_json())
     if not isinstance(value, list):
         msg = "Rust core list_oauth_credentials_json returned non-list JSON"
         raise TypeError(msg)
@@ -195,7 +195,7 @@ def list_oauth_credentials() -> list[dict[str, Any]]:
 
 def clear_oauth_credentials(target: str | None = None) -> list[str]:
     """Clear remembered Rust OAuth credential stores."""
-    value = json.loads(_mcp_compressor_core.clear_oauth_credentials_json(target))
+    value = json.loads(_native.clear_oauth_credentials_json(target))
     if not isinstance(value, list):
         msg = "Rust core clear_oauth_credentials_json returned non-list JSON"
         raise TypeError(msg)
