+++ tests/test_atlassian_mcp_real_world.py
@@ -311,7 +311,7 @@ async def test_atlassian_mcp_config_multi_server_and_streamable_http_port() -> N
 
 def test_atlassian_python_high_level_compressor_client() -> None:
     sys.path.insert(0, str(ROOT / "python" / "mcp-compressor-rust"))
-    rust_package = cast("Any", importlib.import_module("mcp_compressor_rust"))
+    rust_package = cast("Any", importlib.import_module("mcp_compressor"))
 
     client = rust_package.CompressorClient(
         servers={
@@ -334,7 +334,7 @@ def test_atlassian_python_high_level_compressor_client() -> None:
 
 def test_atlassian_python_high_level_generated_clients(tmp_path) -> None:
     sys.path.insert(0, str(ROOT / "python" / "mcp-compressor-rust"))
-    rust_package = cast("Any", importlib.import_module("mcp_compressor_rust"))
+    rust_package = cast("Any", importlib.import_module("mcp_compressor"))
 
     client = rust_package.CompressorClient(
         servers={
@@ -380,7 +380,7 @@ def test_atlassian_python_high_level_generated_clients(tmp_path) -> None:
 
 def test_atlassian_python_native_session() -> None:
     sys.path.insert(0, str(ROOT / "python" / "mcp-compressor-rust"))
-    rust_package = cast("Any", importlib.import_module("mcp_compressor_rust"))
+    rust_package = cast("Any", importlib.import_module("mcp_compressor"))
 
     session = rust_package.start_compressed_session(
         rust_package.CompressedSessionConfig(
